Removed --mode option (now always runs in "files" mode)
The processing mode option has been removed due to the change in rule
format. As the testing is now split between paths and file content it
does not make sense to process each file in turn for each rule in turn,
which is what "rules" mode did. The only processing mode available now
processes each rule in turn for each file, meaning that files that need
to be read are only read once, which is the fastest possible method.

 - Modified --help message: removed "--mode" references and added an
   example.