Initial support for file-rule exceptions
A file called "exceptions.json" can now be placed in the application's
root directory to control which rules will be excluded for which files.
A system to manage this file from scan results will be created, but
until then this file needs to be created and modified by hand.

exceptions.json should contain a map of file path (string) keys with an
array of excluded rule IDs (strings) for that file. For example: -

{
  "/path/to/a/file": [
    "rule:id:1", "rule:id:2"
  ]
}

If the above appears in exceptions.json then rules "rule:id:1" and
"rule:id:2" will NOT be processed for the file /path/to/a/file.

 - Added loadExceptions() function.
 - executeTests(): exceptions are loaded first using loadExceptions()
   and then the array of rules for each file is passed to
   processRulesOnFile().
 - processRulesOnFile(): list of rules is filtered based on the passed
   exceptions array.