Added script to produce exceptions from results
The new script (results_to_exceptions.js) reads a JSON file with results
produced from the Web Exploit Detector and optionally an exceptions.json
file and then prompts the user to choose whether to exclude each matched
file from each rule in turn. Users can choose either to exclude or not
to exclude the file, or to skip the remainder of the files in the
current rule. At the end of processing, a new exceptions.json will be
written.

To un-exclude a file for a rule, the user will need to remove it by
editing exclusions.json directly.

 - src/prompt.js: new module which is used to prompt users for input
   from stdin.
 - package.json: modified and added new testing scripts.