Initial commit on the "babel" branch
The project structure has been modified so that source code (in ES6) is
now compiled (by Babel) to plain JS in the new "./lib" directory. The
index.js and results_to_exceptions.js scripts in the project root
directory now load all code from ./lib, allowing older versions of
Node.js to run the scripts correctly.

The ./lib directory is included in version control so that users can
pull a built version of the project which can be run immediately without
having to install dev-dependencies and run the build tool-chain.

 - Added "babel" and "babel-preset-env" dev-dependencies.
 - Added "build" and "watch:build" npm scripts.
 - Modified ESLint configuration slightly.
 - Added LICENSE file.
 - index.js and results_to_exceptions.js: moved all code to modules
   (cli.js and results_to_exceptions.js respectively) which both contain
   execute() functions that are exported as the module default. The
   root directory scripts now simply import these modules (from ./lib)
   and call execute().
 - Modified test specs so that they import from ./lib.