Snapshots now contain timestamp, webroot and relative paths
The new snapshot format is: -
  - meta: snapshot metadata: -
    - timestamp: Unix time when the snapshot was generated.
    - webroot: the base path used when generating this snapshot.
  - data: array of {file, hash} objects as before, but now the file
    paths are relative to meta.webroot.

The code for comparing snapshots still works with snapshot files in the
previous format, however the --webroot option will need to be specified.
This option is no longer required when working with snapshots in the new
format, unless the webroot value needs to be overridden.